+++ test/spec/import/ModelWiringSpec.js
@@ -40,7 +40,7 @@ describe('import - model wiring', function() {
       expect(startEventShape.type).to.equal('bpmn:StartEvent');
       expect(startEventShape.parent).to.eql(subProcessShape);
 
-      expect(subProcessShape.children.length).to.equal(5);
+      expect(subProcessShape.children.length).to.equal(4);
     }));